pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mfc-control"
version = "0.1.0"
description = "Control system for Bronkhorst Mass Flow Controllers in combustion experiments"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Combustion Lab"}
]
keywords = ["mfc", "mass-flow-controller", "bronkhorst", "combustion"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "numpy>=1.24.0",
    "bronkhorst-propar>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
gui = [
    "tkinter-tooltip>=2.0.0",
]
calibration = [
    "scipy>=1.10.0",
    "pandas>=2.0.0",
]

[project.scripts]
mfc-cli = "mfc_control.cli.main:main"

[project.urls]
Repository = "https://github.com/YOUR_USERNAME/mfc-control"

[tool.hatch.build.targets.wheel]
packages = ["src/mfc_control"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
